shuffle notebook imports around
